WIP: Process .yaml files via external helper
This allows to overcome workgen as frontend and permits a restart of
the format by using schema-based yaml as input. The config helper will
then generate json for the native parser, currently in the existing
format.

If some file not ending on .yaml is passed or stdin is specified via
"-", the new helper is not used, and the existing json format is
expected as input.